+++ src/main/java/com/example/rinha/dto/TransactionResponse.java
@@ -2,5 +2,5 @@ package com.example.rinha.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record TransactionResponse(@JsonProperty("limite") Integer creditLimit, @JsonProperty("saldo") Integer balance) {
+public record TransactionResponse(@JsonProperty("limite") Integer creditLimit, @JsonProperty("saldo") Long balance) {
 }
